+++ src/main/java/kittify/common/registry/RegistryCapability.java
@@ -1,0 +1,10 @@
+package kittify.common.registry;
+
+import kittify.Kittify;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
+
+@Mod.EventBusSubscriber(modid = Kittify.MOD_ID)
+public class RegistryCapability {
+
+}