+++ src/main/java/kittify/core/KittifyCore.java
@@ -33,6 +33,7 @@ public class KittifyCore implements IFMLLoadingPlugin {
 
     @Override
     public void injectData(Map<String, Object> data) {
+        CoreUtil.OBFUSCATED = ((boolean) data.get("runtimeDeobfuscationEnabled"));
     }
 
     @Override